refactor(gfql): move call subsystem under call package
