+++ graphistry/compute/gfql/call/__init__.py
@@ -1,0 +1,11 @@
+"""GFQL method-call execution and validation helpers."""
+
+from graphistry.compute.gfql.call.executor import _thread_local, execute_call
+from graphistry.compute.gfql.call.validation import SAFELIST_V1, validate_call_params
+
+__all__ = [
+    "_thread_local",
+    "execute_call",
+    "SAFELIST_V1",
+    "validate_call_params",
+]
