+++ graphistry/tests/compute/gfql/layouts/test_call_layout_radial.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 
 from graphistry.Engine import Engine
-from graphistry.compute.gfql.call_executor import execute_call
+from graphistry.compute.gfql.call.executor import execute_call
 from graphistry.compute.exceptions import ErrorCode, GFQLTypeError
 from graphistry.tests.test_compute import CGFull
